requirements.txt
Flask
pyodbc
pycryptodome
python-dotenv
flask-cors